chore: Add tag to build front

+++ build.sh
@@ -42,13 +42,37 @@ clean () {
 }
 
 buildNode () {
-  case `uname -s` in
-    MINGW*)
-      docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install --no-bin-links && npm update entcore && node_modules/gulp/bin/gulp.js build" 
-      ;;
-    *)
-      docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install && npm update entcore && node_modules/gulp/bin/gulp.js build --springboard=/home/node/$SPRINGBOARD"
-  esac
+  #jenkins
+  echo "[buildNode] Get branch name from jenkins env..."
+  BRANCH_NAME=`echo $GIT_BRANCH | sed -e "s|origin/||g"`
+  if [ "$BRANCH_NAME" = "" ]; then
+    echo "[buildNode] Get branch name from git..."
+    BRANCH_NAME=`git branch | sed -n -e "s/^\* \(.*\)/\1/p"`
+  fi
+  if [ "$BRANCH_NAME" = "" ]; then
+    echo "[buildNode] Branch name should not be empty!"
+    exit -1
+  fi
+
+  if [ "$BRANCH_NAME" = 'master' ]; then
+      echo "[buildNode] Use entcore version from package.json ($BRANCH_NAME)"
+      case `uname -s` in
+        MINGW*)
+          docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install --no-bin-links && npm update entcore && node_modules/gulp/bin/gulp.js build"
+          ;;
+        *)
+          docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install && npm update entcore && node_modules/gulp/bin/gulp.js build --springboard=/home/node/$SPRINGBOARD"
+      esac
+  else
+      echo "[buildNode] Use entcore tag $BRANCH_NAME"
+      case `uname -s` in
+        MINGW*)
+          docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install --no-bin-links && npm rm --no-save entcore && npm install --no-save entcore@$BRANCH_NAME && node_modules/gulp/bin/gulp.js build"
+          ;;
+        *)
+          docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install && npm rm --no-save entcore && npm install --no-save entcore@$BRANCH_NAME && node_modules/gulp/bin/gulp.js build --springboard=/home/node/$SPRINGBOARD"
+      esac
+  fi
 }
 
 buildGradle () {
